Moved mock API data to /tests; added tests for percentage
