+++ src/js/models/v2/aboutTheData/BaseAgencyRow.js
@@ -18,8 +18,8 @@ BaseAgencyRow.populate = function populate(data) {
         : this._name;
     this.certified = data.recent_publication_date_certified || false;
     // eslint-disable-next-line camelcase
-    this._federalTotal = data.federalTotal?.total_budgetary_resources;
-    this.percentageOfTotalFederalBudget = calculatePercentage(this._budgetAuthority, this._federalTotal, 2);
+    this._federalTotal = data?.federalTotal?.total_budgetary_resources;
+    this.percentageOfTotalFederalBudget = calculatePercentage(this._budgetAuthority, this._federalTotal, '--', 2);
 };
 
 export default BaseAgencyRow;
