+++ tests/components/aboutTheData/AboutTheDataPage-test.jsx
@@ -8,7 +8,7 @@ import * as accountHelpers from 'helpers/accountHelper';
 import * as helpers from "containers/account/WithLatestFy";
 import * as glossaryHelpers from 'helpers/glossaryHelper';
 import * as aboutTheDataHelpers from 'helpers/aboutTheDataHelper';
-import { mockAPI } from 'containers/aboutTheData/AgencyTableMapping';
+import { mockAPI } from '../../containers/aboutTheData/mockData';
 import { mockSubmissions } from '../../mockData/helpers/aboutTheDataHelper';
 
 // latest fy of 2020; latest period is 12
